# prototype/1. Technical_ EU STR - Code Single Digital Entry Point/str-ap-internal-0.0.49 (update February 2025)/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A gateway for the electronic transmission of data between online short-term rental platforms and competent authorities, ensuring timely, reliable and efficient data sharing processes\nPriority of development is: 1. listings, 2. orders, 3. activity, 4. area\n\u003cbr\u003e\nTo obtain API credentials, please contact: wouter.travers@pwc.com via e-mail",
        "title": "EU STR - Single Digital Entry Point",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Wouter Travers",
            "email": "wouter.travers@pwc.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "basePath": "/api/v0",
    "paths": {
        "/ca/activity-data": {
            "get": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Retrieve activity data submitted to the SDEP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ca"
                ],
                "summary": "Retrieve activity data submitted to the SDEP",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of records to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.SingleActivityData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid limit parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized\"\t\t//\t(Handled by OAuth2 middleware)",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "429": {
                        "description": "Too many requests\"\t//\t(Could be enforced by rate-limiting middleware)",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
        "/str/activity-data": {
            "post": {
                "security": [
                    {
                        "OAuth2AccessCode": [
                            "read"
                        ]
                    }
                ],
                "description": "Submit activity data to the SDEP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "str"
                ],
                "summary": "Submit activity data to the SDEP",
                "parameters": [
                    {
                        "description": "json formatted ActivityData",
                        "name": "ActivityData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ActivityData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delivered",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.ActivityData"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Missing mandatory parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized\"\t//\t(Handled by OAuth2 middleware)",
                        "schema": {
                            "$ref": "#/definitions/handler.UnauthorizedError"
                        }
                    },
                    "422": {
                        "description": "Invalid data field values",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "503": {
                        "description": "Service unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ActivityData": {
            "type": "object",
            "required": [
                "data",
                "metadata"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.GuestData"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/handler.Metadata"
                }
            }
        },
        "handler.Address": {
            "type": "object",
            "properties": {
                "city": {
                    "description": "City name",
                    "type": "string",
                    "example": "Diegem"
                },
                "country": {
                    "description": "Country code with ISO 3166-1 alpha-3 validation",
                    "type": "string",
                    "example": "BEL"
                },
                "postalCode": {
                    "description": "Postal code",
                    "type": "string",
                    "example": "1831"
                },
                "street": {
                    "description": "Street address",
                    "type": "string",
                    "example": "Culliganlaan 5"
                }
            }
        },
        "handler.BadRequestError": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "Error message for bad requests",
                    "type": "string",
                    "example": "Wrong data format!"
                }
            }
        },
...
        "handler.GuestData": {
            "type": "object",
            "properties": {
                "URL": {
                    "type": "string",
                    "example": "placeholder-URL"
                },
                "address": {
                    "$ref": "#/definitions/handler.Address"
                },
                "competentAuthorityId_area": {
                    "type": "string",
                    "example": "competentAuthorityId_area"
                },
                "competentAuthorityName_area": {
                    "type": "string",
                    "example": "competentAuthorityName_area"
                },
                "countryOfGuests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "ITA",
                        "NLD"
                    ]
                },
                "hostId": {
                    "type": "string",
                    "example": "placeholder-host-id"
                },
                "numberOfGuests": {
                    "type": "integer",
                    "example": 3
                },
                "registrationNumber": {
                    "type": "string",
                    "example": "placeholder-registrationNumber"
                },
                "temporal": {
                    "$ref": "#/definitions/handler.Temporal"
                }
            }
        },
...
        "handler.InternalServerError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message for internal server errors",
                    "type": "string",
                    "example": "An unexpected error occurred"
                }
            }
        },
...
        "handler.Metadata": {
            "type": "object",
            "properties": {
                "additionalProp1": {
                    "description": "Placeholder for additional properties",
                    "type": "object"
                },
                "platform": {
                    "description": "Platform name",
                    "type": "string",
                    "example": "booking.com"
                },
                "submissionDate": {
                    "description": "Date of submission",
                    "type": "string",
                    "example": "2024-07-21T17:32:28Z"
                }
            }
        },
...
        "handler.SingleActivityData": {
            "type": "object",
            "required": [
                "data",
                "metadata"
            ],
            "properties": {
                "competentAuthorityId_validated": {
                    "type": "string"
                },
                "competentAuthorityName_validated": {
                    "type": "string"
                },
                "data": {
                    "$ref": "#/definitions/handler.GuestData"
                },
                "metadata": {
                    "$ref": "#/definitions/handler.Metadata"
                }
            }
        },
...
        "handler.Temporal": {
            "type": "object",
            "properties": {
                "endDateTime": {
                    "type": "string",
                    "example": "2024-07-25T17:32:28Z"
                },
                "startDateTime": {
                    "type": "string",
                    "example": "2024-07-21T17:32:28Z"
                }
            }
        },
        "handler.UnauthorizedError": {
            "type": "object",
            "properties": {
                "message": {
                    "description": "Error message for unauthorized access",
                    "type": "string",
                    "example": "JWT is invalid"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://tt-dp-dev.eu.auth0.com/authorize?audience=https://str.eu",
            "tokenUrl": "https://tt-dp-dev.eu.auth0.com/oauth/token",
            "scopes": {
                "str": "Grants read access"
            }
        }
    },
...
}